[Controller Skin] Adds optional `placement` field to determine where items/screens should be placed
With Split View controller skins, we want to be able to specify whether a given item (or screen) should be placed in the keyboard game controller or in the app window itself. Now, you can specify `app` or `controller` for the optional `placement` field, and Delta will position the item/screen in the correct area.

NOTE: `app` placement items/screens use normalized [0-1] frames for positioning.